refactor: split author and publication modules
